requirements.txt
asgiref==3.3.4
Django==3.2.2
pip==21.1.1
pytz==2021.1
setuptools==56.2.0
sqlparse==0.4.1
djangorestframework
requests
gunicorn